+++ src/service/aggregate.ts
@@ -157,6 +157,9 @@ function aggregateByEvent(params: {
             // 入場数の集計を行う
             const checkinCountAggregation = aggregateCheckinCount(checkGates, reservations, offers);
 
+            // イベントステータス最終更新時の予約について未入場数を算出する
+            const { uncheckedReservations } = aggregateUncheckedReservations(reservations, event);
+
             aggregation = {
                 id: performance.id,
                 maximumAttendeeCapacity: <number>maximumAttendeeCapacity,
@@ -177,7 +180,7 @@ function aggregateByEvent(params: {
             debug('aggregated!', aggregation);
 
             // パフォーマンスリポジトリにも保管
-            await saveAggregation2performance(aggregation)(repos);
+            await saveAggregation2performance(aggregation, uncheckedReservations)(repos);
         } catch (error) {
             // tslint:disable-next-line:no-console
             console.error('couldn\'t create aggregation on event', performance.id, error);
@@ -188,7 +191,10 @@ function aggregateByEvent(params: {
 /**
  * パフォーマンスコレクションに集計データを保管する
  */
-function saveAggregation2performance(params: factory.performance.IPerformanceAggregation) {
+function saveAggregation2performance(
+    params: factory.performance.IPerformanceAggregation,
+    uncheckedReservations: factory.reservation.event.IReservation[]
+) {
     return async (repos: {
         performance: repository.Performance;
     }) => {
@@ -210,7 +216,14 @@ function saveAggregation2performance(params: factory.performance.IPerformanceAgg
                     : undefined,
                 ...(typeof params.remainingAttendeeCapacityForWheelchair === 'number')
                     ? { remainingAttendeeCapacityForWheelchair: params.remainingAttendeeCapacityForWheelchair }
-                    : undefined
+                    : undefined,
+                'ttts_extension.uncheckedReservations': uncheckedReservations.map((r) => {
+                    return {
+                        typeOf: r.typeOf,
+                        id: r.id,
+                        underName: r.underName
+                    };
+                })
             },
             $unset: {
                 noExistingAttributeName: 1 // $unsetは空だとエラーになるので
@@ -299,5 +312,35 @@ function aggregateCheckinCount(
     return {
         checkinCount: allUniqueCheckins.length,
         checkinCountsByWhere: checkinCountsByWhere
+    };
+}
+
+function aggregateUncheckedReservations(
+    reservations: factory.reservation.event.IReservation[],
+    event: factory.performance.IPerformance
+) {
+    let uncheckedReservations: factory.reservation.event.IReservation[] = [];
+
+    const extension = event.ttts_extension;
+
+    // 時点での予約
+    const reservationsAtLastUpdateDate = extension?.reservationsAtLastUpdateDate;
+    if (Array.isArray(reservationsAtLastUpdateDate)) {
+        // 時点での予約が存在していれば、そのうちの未入場数を検索
+        if (reservationsAtLastUpdateDate.length > 0) {
+            const targetReservationIds = reservationsAtLastUpdateDate.map((r) => r.id);
+            uncheckedReservations = reservations
+                .filter((r) => targetReservationIds.includes(r.id))
+                .filter((r) => r.checkins.length === 0);
+            // uncheckedReservations = await repos.re.search({
+            //     typeOf: factory.chevre.reservationType.EventReservation,
+            //     ids: targetReservationIds,
+            //     checkins: { $size: 0 } // $sizeが0より大きい、という検索は現時点ではMongoDBが得意ではない
+            // });
+        }
+    }
+
+    return {
+        uncheckedReservations
     };
 }