+++ src/credentials.ts
@@ -11,8 +11,5 @@ export const credentials = {
     lineNotify: {
         url: <string>process.env.LINE_NOTIFY_URL,
         accessToken: <string>process.env.LINE_NOTIFY_ACCESS_TOKEN
-    },
-    sendGrid: {
-        apiKey: <string>process.env.SENDGRID_API_KEY
     }
 };